requirements.txt
functions-framework
google-cloud-logging
slackclient
Werkzeug
